requirements.txt
flask==0.11.1
gunicorn==19.7.1
google.cloud.datastore==1.6.0
google-cloud-translate==1.3.1
google-cloud-language==1.0.2